requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
joblib
pyyaml
earthengine-api
